LLM-service/requirements.txt
fastapi
uvicorn
httpx
python-multipart
pydantic
python-dotenv
torch
transformers
langchain
langchain-huggingface
